requirements.txt
faiss-cpu
sentence-transformers
datasets
tqdm
requests
transformers
torch
